+++ percy/parser/__init__.py
@@ -1,1 +1,0 @@
-# This package contains supporting files for the `RecipeParser` class