+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 Note: version releases in the 0.x.y range may introduce breaking changes.
 
+## 0.2.0
+- Remove previous parser tree code and depend on conda-recipe-manager instead.
+- Remove `get_read_only_parser()` function.
+
 ## 0.1.7
 - Revert exclusion of `clang-compiler-activation-feedstock` from locally loaded feedstocks since it's actually active and needed.